Move paths functions for active version to separate file
